requirements.txt
flask==3.0.0
gunicorn==23.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
tabulate==0.9.0
python-dotenv==1.0.0
# Add fogis_api_client as a dependency with specific version
fogis-api-client-timmyBird==0.0.5
# Logging and monitoring dependencies
structlog==24.1.0
python-json-logger==2.0.7
prometheus-flask-exporter==0.23.0
